RAG/requirements.txt
transformers
openai
faiss-cpu
sentence_transformers
flask
flask_cors
gunicorn
python-dotenv